requirements.txt
tornado
mysql-connector-python
sqlalchemy
sockjs-tornado
redis
wtforms
werkzeug